moved files to allow direct access to skeleton
